Moved LocationCapability and LocationOperation into their own podspec
