+++ PSOperationsLocation/LocationCapability-tvOS.swift
@@ -1,21 +1,22 @@
 //
-//  LocationCapability-OSX.swift
+//  LocationCapability-tvOS.swift
 //  PSOperations
 //
 //  Created by Dev Team on 10/4/15.
 //  Copyright © 2015 Pluralsight. All rights reserved.
 //
 
-#if os(OSX)
+#if os(tvOS)
 
 import Foundation
 import CoreLocation
+import PSOperations
 
 public struct Location: CapabilityType {
     public static let name = "Location"
 
     public init() { }
-    
+
     public func requestStatus(_ completion: @escaping (CapabilityStatus) -> Void) {
         guard CLLocationManager.locationServicesEnabled() else {
             completion(.notAvailable)
@@ -25,16 +26,12 @@ public struct Location: CapabilityType {
         let actual = CLLocationManager.authorizationStatus()
         
         switch actual {
-        case .notDetermined:
-            completion(.notDetermined)
-        case .restricted:
-            completion(.notAvailable)
-        case .denied:
-            completion(.denied)
-        case .authorized:
-            completion(.authorized)
-        default:
-            completion(.authorized)
+            case .notDetermined: completion(.notDetermined)
+            case .restricted: completion(.notAvailable)
+            case .denied: completion(.denied)
+            case .authorizedWhenInUse: completion(.authorized)
+            case .authorizedAlways:
+                fatalError(".Always should be unavailable on tvOS")
         }
     }
     
@@ -45,7 +42,7 @@ public struct Location: CapabilityType {
 
 private let Authorizer = LocationAuthorizer()
 
-fileprivate class LocationAuthorizer: NSObject, CLLocationManagerDelegate {
+private class LocationAuthorizer: NSObject, CLLocationManagerDelegate {
     
     private let manager = CLLocationManager()
     private var completion: ((CapabilityStatus) -> Void)?
@@ -61,31 +58,28 @@ fileprivate class LocationAuthorizer: NSObject, CLLocationManagerDelegate {
         }
         self.completion = completion
         
-        let key = "NSLocationUsageDescription"
-        manager.startUpdatingLocation()
+        let key = "NSLocationWhenInUseUsageDescription"
+        manager.requestWhenInUseAuthorization()
         
         // This is helpful when developing an app.
         assert(Bundle.main.object(forInfoDictionaryKey: key) != nil, "Requesting location permission requires the \(key) key in your Info.plist")
     }
     
-    
     @objc func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        if let completion = self.completion, manager == self.manager {
+        if let completion = self.completion, manager == self.manager && status != .notDetermined {
             self.completion = nil
             
             switch status {
-            case .authorized:
-                completion(.authorized)
-            case .denied:
-                completion(.denied)
-            case .restricted:
-                completion(.notAvailable)
-            case .notDetermined:
-                self.completion = completion
-                manager.startUpdatingLocation()
-                manager.stopUpdatingLocation()
-            default:
-                completion(.authorized)
+                case .authorizedWhenInUse:
+                    completion(.authorized)
+                case .denied:
+                    completion(.denied)
+                case .restricted:
+                    completion(.notAvailable)
+                case .authorizedAlways:
+                    fatalError(".Always should be unavailable on tvOS")
+                case .notDetermined:
+                    fatalError("Unreachable due to the if statement, but included to keep clang happy")
             }
         }
     }
